Inventory is done
Was able to update db when new customer purchase
was able to update the amount in warhouse after purchase
ToDO: work on editInventory / and Change settings to history purchases ie loading customers

+++ src/database/InventoryFindOne.java
@@ -12,7 +12,7 @@ import com.sun.xml.internal.txw2.Document;
 
 public class InventoryFindOne {
 	
-	public InventoryFindOne (DefaultTableModel tableModel,MongoDatabase db,int barcode) {
+	public InventoryFindOne (DefaultTableModel tableModel,MongoDatabase db,Long barcode) {
 		MongoCollection<org.bson.Document> collection = db.getCollection("Inventory");
 //		MongoCursor<org.bson.Document> cursor = collection.find().iterator();
 		Bson bsonFilter = Filters.eq("Barcode", barcode);
